website-bundler/pyproject.toml
[tool.poetry]
name = "kioskC2"
version = "0.0.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
nats-py = "^2.13.1"
pydantic = "^2.12.5"
playwright = "^1.58.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"